+++ src/components/TextSection/index.jsx
@@ -6,8 +6,24 @@ import Reserva from "../Reserva/index";
 import Derechos from "../Derechos";
 import SectionParagraphPoliticals from "../SectionParagraphPoliticals";
 
+const getComponent = (type) =>
+  type === "Cokkies"
+    ? Cokkies
+    : type === "Licencia"
+    ? Licencia
+    : type === "Hipervinculos"
+    ? Hipervinculos
+    : type === "Responsabilidad"
+    ? Responsabilidad
+    : type === "Reserva"
+    ? Reserva
+    : type === "Derechos"
+    ? Derechos
+    : SectionParagraphPoliticals;
+
 const TextSection = ({
   title,
+  type,
   id,
   allParagraph,
   allParagraph2,
@@ -25,9 +41,11 @@ const TextSection = ({
   allList = allList && Object.values(allList);
   allList2 = allList2 && Object.values(allList2);
   allList3 = allList3 && Object.values(allList3);
-  if (title === "Cokkies:") {
+  const Component = getComponent(type);
+
+  if (Component) {
     textSection = (
-      <Cokkies
+      <Component
         title={title}
         allParagraph={allParagraph}
         allParagraph2={allParagraph2}
@@ -39,88 +57,6 @@ const TextSection = ({
         id={id}
       />
     );
-  } else {
-    if (title === "Licencia:") {
-      textSection = (
-        <Licencia
-          title={title}
-          allParagraph={allParagraph}
-          allParagraph2={allParagraph2}
-          allParagraph3={allParagraph3}
-          allParagraph4={allParagraph4}
-          allList={allList}
-          allList2={allList2}
-          allList3={allList3}
-          id={id}
-        />
-      );
-    } else {
-      if (title === "Hipervínculos a nuestro contenido:") {
-        textSection = (
-          <Hipervínculos
-            title={title}
-            allParagraph={allParagraph}
-            allParagraph2={allParagraph2}
-            allParagraph3={allParagraph3}
-            allParagraph4={allParagraph4}
-            allList={allList}
-            allList2={allList2}
-            allList3={allList3}
-            id={id}
-          />
-        );
-      } else {
-        if (title === "Responsabilidad del contenido:") {
-          textSection = (
-            <Responsabilidad
-              title={title}
-              allParagraph={allParagraph}
-              allParagraph2={allParagraph2}
-              allParagraph3={allParagraph3}
-              allParagraph4={allParagraph4}
-              allList={allList}
-              allList2={allList2}
-              allList3={allList3}
-              id={id}
-            />
-          );
-        } else {
-          if (title === "Reserva de derechos:") {
-            textSection = (
-              <Reserva
-                title={title}
-                allParagraph={allParagraph}
-                allParagraph2={allParagraph2}
-                allParagraph3={allParagraph3}
-                allParagraph4={allParagraph4}
-                allList={allList}
-                allList2={allList2}
-                allList3={allList3}
-                id={id}
-              />
-            );
-          } else {
-            if (title === "Tus derechos:") {
-              textSection = (
-                <Derechos
-                  allParagraph={allParagraph}
-                  title={title}
-                  allList={allList}
-                  allParagraph2={allParagraph2}
-                />
-              );
-            } else {
-              textSection = (
-                <SectionParagraphPoliticals
-                  allParagraph={allParagraph}
-                  title={title}
-                />
-              );
-            }
-          }
-        }
-      }
-    }
   }
 
   return <>{textSection}</>;
