+++ src/pages/politicas/index.jsx
@@ -1,65 +1,7 @@
 import styles from "./styles.module.css";
 import BlockStyle from "@/components/BlockStyle/BlockStyle";
 import TextSection from "@/components/TextSection/index";
-
-const alltext = [
-  {
-    title: "Información personal que recopilamos:",
-    allParagraph: {
-      0: 'Cuando visitas Face It! Citas Rápidas, recopilamos automáticamente cierta información sobre tu dispositivo, incluida información sobre tu navegador web, dirección IP, zona horaria y algunas de las cookies instaladas en tu dispositivo. Además, a medida que navegas, recopilamos información sobre las páginas web individuales o los productos que ves, qué sitios web o términos de búsqueda te remitieron a la web y cómo interactúas. Nos referimos a esta información recopilada automáticamente como "Información del dispositivo". Además, podemos recopilar los datos personales que nos proporcionas (incluidos, entre otros, nombre, apellido, dirección, información de pago, etc.) durante el registro para poder cumplir con el acuerdo.',
-    },
-  },
-  {
-    title: "¿Por qué procesamos tus datos?",
-    allParagraph: {
-      0: "Puedes visitar la web sin decirnos quién eres ni revelar ninguna información por la cual alguien pueda identificarte como una persona específica. Sin embargo, si deseas utilizar algunas de las funciones del sitio web, o deseas recibir nuestro boletín informativo o proporcionar otros detalles al completar un formulario, puedes proporcionarnos datos personales, como tu correo electrónico, nombre, apellido, ciudad de residencia, organización y número de teléfono. Puedes optar por no proporcionar tus datos personales, pero es posible que no puedas aprovechar algunas de las funciones del sitio web. Por ejemplo, no podrás recibir nuestro boletín ni contactarnos directamente desde el sitio web. Los usuarios que no estén seguros de qué información es obligatoria pueden ponerse en contacto con nosotros a través de info@faceitcitasrapidas.com.",
-    },
-  },
-  {
-    title: "Tus derechos:",
-    allParagraph: {
-      0: "Si eres residente europeo, tienes los siguientes derechos relacionados con tus datos personales:",
-    },
-    allList: {
-      0: "El derecho a ser informado.",
-      1: "El derecho de acceso.",
-      2: "El derecho a la rectificación.",
-      3: "El derecho a borrar.",
-      4: "El derecho a restringir el procesamiento.",
-      5: "El derecho a la portabilidad de datos.",
-      6: "El derecho a oponerte.",
-      7: "Derechos en relación con la toma de decisiones automatizada y la elaboración de perfiles.",
-    },
-    allParagraph2: {
-      0: "Si deseas ejercer este derecho, comunícate con nosotros a través de la información de contacto a continuación.",
-      1: "Además, si eres residente europeo, destacamos que estamos procesando tu información para cumplir con los contratos que podríamos tener contigo (por ejemplo, si realizas un pedido a través de la web), o de otra manera para seguir nuestros intereses comerciales legítimos enumerados anteriormente. Además, ten en cuenta que tu información puede transferirse fuera de Europa, incluidos Canadá y Estados Unidos.",
-    },
-  },
-  {
-    title: "Enlaces a otros sitios web:",
-    allParagraph: {
-      0: "Nuestra web puede contener enlaces a otros sitios web que no son de nuestra propiedad ni están controlados por nosotros. Ten en cuenta que no somos responsables de dichos sitios web ni de las prácticas de privacidad de terceros. Te recomendamos que estés atento cuando abandones nuestro sitio web y leas las declaraciones de privacidad de cada web que pueda recopilar información personal.",
-    },
-  },
-  {
-    title: "Seguridad de la información:",
-    allParagraph: {
-      0: "Aseguramos la información que proporcionas en servidores informáticos en un entorno controlado y seguro, protegido del acceso, uso o divulgación no autorizados. Mantenemos medidas de seguridad administrativas, técnicas y físicas razonables para proteger contra el acceso no autorizado, el uso, la modificación y la divulgación de datos personales bajo su control y custodia. Sin embargo, no se puede garantizar la transmisión de datos a través de Internet o redes inalámbricas.",
-    },
-  },
-  {
-    title: "Divulgación legal:",
-    allParagraph: {
-      0: "Divulgaremos cualquier información que recopilemos, usemos o recibamos si así lo requiere o lo permite la ley, como para cumplir con una citación o un proceso legal similar, y cuando creemos de buena fe que la divulgación es necesaria para proteger nuestros derechos, proteger tu seguridad o la seguridad de los demás, investigar el fraude o responder a una solicitud del gobierno.",
-    },
-  },
-  {
-    title: "Información de contacto:",
-    allParagraph: {
-      0: "Si deseas comunicarte con nosotros para comprender más sobre esta Política o deseas comunicarte con nosotros en relación con cualquier asunto sobre los derechos individuales y tu información personal, puedes enviarnos un correo electrónico a info@faceitcitasrapidas.com.",
-    },
-  },
-];
+import { alltext } from "./string";
 
 const PolitcasPrivacidad = () => {
   return (
@@ -86,6 +28,7 @@ const PolitcasPrivacidad = () => {
           {alltext.map((section, key) => (
             <TextSection
               key={key}
+              type={section.type}
               id={section.id}
               title={section.title}
               allParagraph={section.allParagraph}
